long

+++ src/data_collect_modi.h
@@ -1,8 +1,8 @@
-int data_collect_modi(int tid, int ns, int grid_y, int thread, float *gather, float *data, int *rcr_position_collect, int n, int n_src, int n_rcr)
+int data_collect_modi(long long int  tid, long long int  ns, long long int  grid_y, long long int  thread, float *gather, float *data, long long int  *rcr_position_collect, long long int  n, long long int  n_src, long long int  n_rcr)
 {
-    for(int i = 0; i < ns; i++)
+    for(long long int  i = 0; i < ns; i++)
     {
-        for(int j = 0; j < n_rcr; j++)
+        for(long long int  j = 0; j < n_rcr; j++)
         {
             data[(tid * ns * n_rcr) + n_rcr * i + j] = gather[tid * ns * grid_y + rcr_position_collect[(n) * thread + (tid + n_src * j)] + grid_y * i];
         }
